+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## v6.0.0
+- IFS-2008: Anbindung der HttpEncodingAutoConfiguration
 - IFS-1156: Für Data Table: Farbliche Hervorhebung der in Detail-Ansicht gezeigten Zeile
 - IFS-1946: HttpHeaderNestedDiagnosticContextFilter entfernt, da bereits in Isy-Aufrufkontext definiert
 - IFS-1616: Anpassung an Spring Boot Version 2.7.2